requirements.txt
torch==1.2.0
torchvision==0.4.0
easydict
matplotlib
opencv-python
PyQt5==5.13.0
PyQt5-sip==4.19.18
glob3
pillow
dill
dominate>=2.3.1 
scikit-image
QDarkStyle==2.7
qdarkgraystyle==1.0.2
tensorboard==1.14.0
tensorboardX==1.9
tqdm==4.32.1
urllib3==1.25.8
visdom==0.1.8.9
numpy
sklearn
simple_3dviz
tripy
shapely
scipy
pywavefront